refactor(search): move vector-distance env reader to embeddings/config.ts
getVectorMaxDistance, the DEFAULT_*/CEILING constants, and the warnOnce helper
read process.env, but lived in exact-search.ts which is otherwise pure cosine
math. Move them to embeddings/config.ts alongside the other GITNEXUS_* embedding
readers (resolveEmbeddingConfig, parsePositiveInt), and repoint the imports in
embedding-pipeline.ts, local-backend.ts, and the unit test. exact-search.ts
returns to an import-free pure module. No behavior change.
